+++ src/components/Annotation/ReplyForm/MobileReplyForm.css
@@ -1,0 +1,13 @@
+.mobile-reply-form {
+  height: 100%;
+  left: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+  z-index: 10;
+}
+
+.mobile-reply-form textarea {
+  width: 100%;
+  height: 100%;
+}
