+++ src/components/Annotation/ReplyForm/MobileReplyForm.tsx
@@ -1,0 +1,21 @@
+import { createPortal } from 'react-dom';
+
+import './MobileReplyForm.css';
+
+export const MobileReplyForm = () => {
+
+  return (
+    <div>Reply...</div>
+  )
+
+  /*
+  return createPortal(
+    <div className="mobile-reply-form">
+      <textarea autoFocus/>
+    </div>,
+    
+    document.body
+  );
+  */
+
+}
